refactor(domain): extract urban risk functions into urban_risk.go and slim investment.go
Move BuildUrbanRisksFromAPIs, BuildHazardRisks, calcZoningSummary,
detectUrbanRisks, and modalString from investment.go to urban_risk.go.
investment.go now contains only the orchestration entry point Analyze(),
calcCriticalErrors, CalcDSCR, CalcLandPriceStats, CompareLandPrice,
SqmToTsubo, TsuboToSqm, and package-level constants (254 lines).
