+++ backend/internal/domain/investment.go
@@ -3,10 +3,7 @@ package domain
 import (
 	"context"
 	"fmt"
-	"math"
 	"sort"
-	"strings"
-	"sync"
 
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
@@ -18,348 +15,12 @@ const (
 	// SqmPerTsubo は 1坪あたりの平方メートル数（mlit パッケージからも参照）
 	SqmPerTsubo = 3.30578
 
-	LoanMethodEqualPayment   = "equal-payment"   // 元利均等返済
-	LoanMethodEqualPrincipal = "equal-principal"  // 元金均等返済
+	LoanMethodEqualPayment   = "equal-payment"  // 元利均等返済
+	LoanMethodEqualPrincipal = "equal-principal" // 元金均等返済
 
-	DepreciationMethodStraightLine     = "straight-line"      // 定額法
-	DepreciationMethodDecliningBalance = "declining-balance"  // 定率法
+	DepreciationMethodStraightLine     = "straight-line"     // 定額法
+	DepreciationMethodDecliningBalance = "declining-balance" // 定率法
 )
-
-// resolveRateForYear はスケジュールと基準金利から指定年の適用金利を返す。
-// schedule が空の場合は baseRate + rateDelta を返す（固定金利）。
-// rateDelta はストレステスト用の追加オフセット。
-func resolveRateForYear(baseRate, rateDelta float64, schedule []RateAdjustment, year int) float64 {
-	rate := baseRate
-	for _, adj := range schedule {
-		if year >= adj.AfterYear {
-			rate = adj.Rate
-		}
-	}
-	return rate + rateDelta
-}
-
-// yieldParams は yield / vacancy の初期計算結果をまとめた内部 struct。
-type yieldParams struct {
-	miscExpenses    float64
-	totalInvestment float64
-	annualRent      float64
-	grossYield      float64
-	netYield        float64
-}
-
-// loanParams はローン初期値をまとめた内部 struct。
-// currentRate / monthlyPayment は変動金利スケジュールによりループ内で変化するため、
-// ループ開始時にローカル変数へコピーして使う。
-type loanParams struct {
-	currentRate           float64
-	monthlyPayment        float64
-	monthlyPrincipalFixed float64 // 元金均等返済のみ非ゼロ
-}
-
-// depreciationParams は減価償却初期値をまとめた内部 struct。
-// bookValue は定率法でループ内に変化するため、ループ開始時にローカル変数へコピーして使う。
-type depreciationParams struct {
-	usefulLife         int
-	annualDepreciation float64 // 定額法定数。定率法では 0
-	bookValue          float64 // 定率法のみ非ゼロ
-	decliningRate      float64 // 定率法のみ非ゼロ
-}
-
-// simulationResult は年次シミュレーションループの出力をまとめた内部 struct。
-type simulationResult struct {
-	yearlyResults           []YearlyResult
-	accumulatedDepreciation float64
-	deadCrossYear           int     // -1 = デッドクロスなし
-	totalInterest           float64 // 保有期間の総支払利息
-}
-
-// irrNPVResult は IRR / NPV 計算結果をまとめた内部 struct。
-type irrNPVResult struct {
-	irr *float64
-	npv float64
-}
-
-func initYieldParams(input InvestmentInput) yieldParams {
-	effectiveVacancy := math.Min(input.VacancyRate+input.VacancyRateDelta, 0.99)
-	loanFee := input.LoanAmount * input.LoanFeeRate
-	miscExpenses := (input.LandPrice+input.BuildingCost)*input.MiscExpenseRate + loanFee
-	totalInvestment := input.LandPrice + input.BuildingCost + miscExpenses
-	annualRent := input.MonthlyRent * 12 * (1 - effectiveVacancy)
-	grossYield := 0.0
-	if totalInvestment > 0 {
-		grossYield = (input.MonthlyRent * 12) / totalInvestment
-	}
-	annualExpenses := annualRent * input.ExpenseRate
-	netYield := 0.0
-	if totalInvestment > 0 {
-		netYield = (annualRent - annualExpenses) / totalInvestment
-	}
-	return yieldParams{
-		miscExpenses:    miscExpenses,
-		totalInvestment: totalInvestment,
-		annualRent:      annualRent,
-		grossYield:      grossYield,
-		netYield:        netYield,
-	}
-}
-
-func initLoanParams(input InvestmentInput) loanParams {
-	currentRate := resolveRateForYear(input.AnnualLoanRate, input.LoanRateDelta, input.RateAdjustmentSchedule, 1)
-	var monthlyPayment, monthlyPrincipalFixed float64
-	if input.LoanMethod == LoanMethodEqualPrincipal && input.LoanYears > 0 {
-		monthlyPrincipalFixed = input.LoanAmount / float64(input.LoanYears*12)
-	} else {
-		monthlyPayment = calcMonthlyPayment(input.LoanAmount, currentRate, input.LoanYears)
-	}
-	return loanParams{
-		currentRate:           currentRate,
-		monthlyPayment:        monthlyPayment,
-		monthlyPrincipalFixed: monthlyPrincipalFixed,
-	}
-}
-
-func initDepreciationParams(input InvestmentInput) depreciationParams {
-	// 中古物件は簡便法耐用年数を使用（新築は法定耐用年数）
-	usefulLife := CalcResidualUsefulLife(input.BuildingType, input.BuildingAge)
-	annualDepreciation := input.BuildingCost / float64(usefulLife)
-	// bookValue / decliningRate は定率法のみで使用する（ループをまたいで簿価を追跡）
-	var bookValue, decliningRate float64
-	if input.DepreciationMethod == DepreciationMethodDecliningBalance {
-		bookValue = input.BuildingCost
-		decliningRate = 1.5 / float64(usefulLife)
-	}
-	return depreciationParams{
-		usefulLife:         usefulLife,
-		annualDepreciation: annualDepreciation,
-		bookValue:          bookValue,
-		decliningRate:      decliningRate,
-	}
-}
-
-// capexForYear は CapexSchedule から指定年の修繕費合計を返す。
-func capexForYear(schedule []CapexEvent, year int) float64 {
-	total := 0.0
-	for _, ev := range schedule {
-		if ev.Year == year {
-			total += ev.Amount
-		}
-	}
-	return total
-}
-
-// rentForYear は賃料上昇・下落シナリオを考慮した N 年目の年間実効賃料を返す。
-// y は 0-indexed（1年目 = y=0）。
-//
-// 設計: y=0 は「初年度（変化なし）」を表す（RentDeclineRate の既存挙動と対称）。
-// 上昇期 y=0..rentGrowthYears-1 は (1+g)^y を適用し、y=rentGrowthYears-1 が上昇期最終年。
-// y=rentGrowthYears は「上昇期終了直後の最初の下落年」であり、ピーク × (1-d)^1 を返す。
-// これは "下落1年目が即始まる保守的モデル" を採用したもの。
-// peak は上昇が続いたと仮定した場合の理論値 (1+g)^rentGrowthYears を基準とする。
-func rentForYear(baseRent float64, rentDeclineRate, rentGrowthRate float64, rentGrowthYears, y int) float64 {
-	if rentGrowthRate <= 0 || rentGrowthYears <= 0 {
-		return baseRent * math.Pow(1-rentDeclineRate, float64(y))
-	}
-	if y < rentGrowthYears {
-		return baseRent * math.Pow(1+rentGrowthRate, float64(y))
-	}
-	peak := baseRent * math.Pow(1+rentGrowthRate, float64(rentGrowthYears))
-	return peak * math.Pow(1-rentDeclineRate, float64(y-rentGrowthYears+1))
-}
-
-// simulateYears は年次 P&L ループを実行し yearlyResults・デッドクロス年・累積減価償却額を返す。
-// years は呼び出し元 Analyze() が決定した値をそのまま受け取る。
-func simulateYears(input InvestmentInput, years int, yp yieldParams, lp loanParams, dp depreciationParams) simulationResult {
-	yearlyResults := make([]YearlyResult, years)
-	remainingBalance := input.LoanAmount
-	cumulativeCF := 0.0
-	deadCrossYear := -1
-	var accumulatedDepreciation float64
-	var totalInterest float64
-
-	// loanParams / depreciationParams の可変フィールドをループローカルにコピー
-	currentRate := lp.currentRate
-	monthlyPayment := lp.monthlyPayment
-	bookValue := dp.bookValue
-
-	for y := 0; y < years; y++ {
-		year := y + 1
-
-		// 変動金利: スケジュールで金利が変化したら返済額を残高・残期間で再計算（元利均等のみ）
-		if year > 1 && len(input.RateAdjustmentSchedule) > 0 {
-			newRate := resolveRateForYear(input.AnnualLoanRate, input.LoanRateDelta, input.RateAdjustmentSchedule, year)
-			if newRate != currentRate && remainingBalance > 0 && year <= input.LoanYears {
-				if input.LoanMethod != LoanMethodEqualPrincipal {
-					remainingYears := input.LoanYears - y
-					monthlyPayment = calcMonthlyPayment(remainingBalance, newRate, remainingYears)
-				}
-				currentRate = newRate
-			}
-		}
-
-		annualInterest := 0.0
-		annualPrincipal := 0.0
-		annualLoanPayment := 0.0
-
-		if remainingBalance > 0 && year <= input.LoanYears {
-			if input.LoanMethod == LoanMethodEqualPrincipal {
-				annualInterest, annualPrincipal = calcYearlyLoanComponentsEqualPrincipal(
-					remainingBalance, currentRate, lp.monthlyPrincipalFixed,
-				)
-				annualLoanPayment = annualInterest + annualPrincipal
-			} else {
-				annualInterest, annualPrincipal = calcYearlyLoanComponents(
-					remainingBalance, currentRate, monthlyPayment,
-				)
-				annualLoanPayment = monthlyPayment * 12
-			}
-			remainingBalance -= annualPrincipal
-			if remainingBalance < 0 {
-				remainingBalance = 0
-			}
-		}
-
-		// 保有期間内の利息のみ集計（HoldingYears 以内の年）
-		if year <= input.HoldingYears {
-			totalInterest += annualInterest
-		}
-
-		yearAnnualRent := rentForYear(yp.annualRent, input.RentDeclineRate, input.RentGrowthRate, input.RentGrowthYears, y)
-		yearExpenses := yearAnnualRent*input.ExpenseRate + input.AnnualPropertyTax
-
-		// 減価償却（定額法または定率法）
-		var yearDepreciation float64
-		switch input.DepreciationMethod {
-		case DepreciationMethodDecliningBalance:
-			if bookValue > 1.0 {
-				yearDepreciation = bookValue * dp.decliningRate
-				if bookValue-yearDepreciation < 1.0 {
-					yearDepreciation = bookValue - 1.0
-				}
-				bookValue -= yearDepreciation
-			}
-		default:
-			if year <= dp.usefulLife {
-				yearDepreciation = dp.annualDepreciation
-			}
-		}
-		accumulatedDepreciation += yearDepreciation
-
-		// 課税所得 = 収入 - 利息 - 減価償却 - 経費
-		taxableIncome := yearAnnualRent - annualInterest - yearDepreciation - yearExpenses
-		incomeTax := 0.0
-		if taxableIncome > 0 {
-			incomeTax = taxableIncome * input.IncomeTaxRate
-		}
-
-		capex := capexForYear(input.CapexSchedule, year)
-		cashFlow := yearAnnualRent - annualLoanPayment - yearExpenses - capex
-		afterTaxCF := cashFlow - incomeTax
-		cumulativeCF += afterTaxCF
-
-		// デッドクロス判定: 元金返済額 > 減価償却費 となるゾーン
-		// 耐用年数経過後は減価償却=0のため、元金返済が残っていれば即デッドクロス
-		// ローン完済後（annualPrincipal==0）はデッドクロスゾーンから脱出
-		// 建物費用=0の場合は減価償却対象資産がなくデッドクロスの概念が適用されないため除外
-		inDeadCrossZone := input.BuildingCost > 0 && annualPrincipal > 0 && annualPrincipal > yearDepreciation
-		isDeadCrossYear := false
-		if deadCrossYear == -1 && inDeadCrossZone {
-			deadCrossYear = year
-			isDeadCrossYear = true
-		}
-
-		yearlyResults[y] = YearlyResult{
-			Year:                 year,
-			CapexAmount:          capex,
-			AnnualRent:           yearAnnualRent,
-			AnnualLoanPayment:    annualLoanPayment,
-			AnnualInterest:       annualInterest,
-			AnnualPrincipal:      annualPrincipal,
-			AnnualDepreciation:   yearDepreciation,
-			AnnualExpenses:       yearExpenses,
-			TaxableIncome:        taxableIncome,
-			IncomeTax:            incomeTax,
-			CashFlow:             cashFlow,
-			AfterTaxCashFlow:     afterTaxCF,
-			RemainingLoanBalance: remainingBalance,
-			CumulativeCashFlow:   cumulativeCF,
-			IsDeadCrossYear:      isDeadCrossYear,
-			IsInDeadCrossZone:    inDeadCrossZone,
-			EffectiveRate:        currentRate,
-		}
-	}
-
-	return simulationResult{
-		yearlyResults:           yearlyResults,
-		accumulatedDepreciation: accumulatedDepreciation,
-		deadCrossYear:           deadCrossYear,
-		totalInterest:           totalInterest,
-	}
-}
-
-// calcAllStressScenarios は 6 つのデフォルトシナリオ（入力値が非ゼロなら第 7 カスタムシナリオも）を
-// goroutine で並列計算して返す。インデックス固定の slice に直書きするため mutex 不要。
-func calcAllStressScenarios(ctx context.Context, input InvestmentInput) []StressScenarioResult {
-	type scenarioDef struct {
-		label     string
-		rateDelta float64
-		vacDelta  float64
-	}
-	allScenarioDefs := [7]scenarioDef{
-		{"ベースライン", 0, 0},
-		{"金利+1%", 0.01, 0},
-		{"金利+2%", 0.02, 0},
-		{"空室+10%", 0, 0.10},
-		{"空室+20%", 0, 0.20},
-		{"複合ストレス", 0.02, 0.10},
-		{"カスタム", input.LoanRateDelta, input.VacancyRateDelta},
-	}
-	hasCustom := input.LoanRateDelta != 0 || input.VacancyRateDelta != 0
-	scenarioCount := 6
-	if hasCustom {
-		scenarioCount = 7
-	}
-	scenarioResults := make([]StressScenarioResult, scenarioCount)
-	var wg sync.WaitGroup
-	for i := 0; i < scenarioCount; i++ {
-		wg.Add(1)
-		go func(idx int) {
-			defer wg.Done()
-			sc := allScenarioDefs[idx]
-			scenarioResults[idx] = calcStressScenario(ctx, input, sc.label, sc.rateDelta, sc.vacDelta)
-		}(i)
-	}
-	wg.Wait()
-	return scenarioResults
-}
-
-// calcIRRNPV は IRR / NPV を計算して返す。
-// equity がゼロ以下（オーバーローン）または HoldingYears=0 の場合は計算が成立しないためゼロ値を返す。
-func calcIRRNPV(
-	input InvestmentInput,
-	yearlyResults []YearlyResult,
-	equity float64,
-	exitNet float64,
-	exitSalePrice float64,
-	accumulatedDepreciation float64,
-	miscExpenses float64,
-) irrNPVResult {
-	if equity <= 0 || input.HoldingYears <= 0 {
-		return irrNPVResult{}
-	}
-	irrCFs := make([]float64, input.HoldingYears)
-	for i := 0; i < input.HoldingYears && i < len(yearlyResults); i++ {
-		irrCFs[i] = yearlyResults[i].AfterTaxCashFlow
-	}
-	irrTerminalValue := exitNet
-	if input.PriceDeclineRate > 0 {
-		decayFactor := math.Pow(1-input.PriceDeclineRate, float64(input.HoldingYears))
-		adjustedSalePrice := exitSalePrice * decayFactor
-		irrTerminalValue = calcTerminalValueWithDecline(input, yearlyResults, adjustedSalePrice, accumulatedDepreciation, miscExpenses)
-	}
-	npv := CalcNPV(irrCFs, irrTerminalValue, input.DiscountRate, equity)
-	irr, _ := CalcIRR(irrCFs, irrTerminalValue, equity)
-	return irrNPVResult{irr: irr, npv: npv}
-}
 
 // Analyze は投資入力値から収支シミュレーション結果を算出する
 func Analyze(ctx context.Context, input InvestmentInput) InvestmentResult {
@@ -447,162 +108,6 @@ func Analyze(ctx context.Context, input InvestmentInput) InvestmentResult {
 	}
 }
 
-// calcYieldScenarios は楽観・標準・悲観の3シナリオにおける年間実効賃料と表面利回りを算出する。
-// 楽観: vacancyRate × 0.5、標準: vacancyRate × 1.0、悲観: vacancyRate × 1.5
-// 注意: 表面利回りは満室想定年収/総投資額（空室率に依存しない）であるが、
-//   AnnualRent（実効賃料）は空室率を反映した値を返す。
-func calcYieldScenarios(input InvestmentInput, totalInvestment float64) YieldScenarios {
-	grossYield := 0.0
-	if totalInvestment > 0 {
-		grossYield = (input.MonthlyRent * 12) / totalInvestment
-	}
-
-	calcScenario := func(vacancyMultiplier float64) YieldScenario {
-		effectiveVacancy := math.Min(input.VacancyRate*vacancyMultiplier, 0.99)
-		annualRent := input.MonthlyRent * 12 * (1 - effectiveVacancy)
-		return YieldScenario{
-			AnnualRent: annualRent,
-			GrossYield: grossYield,
-		}
-	}
-
-	return YieldScenarios{
-		Optimistic:  calcScenario(0.5),
-		Standard:    calcScenario(1.0),
-		Pessimistic: calcScenario(1.5),
-	}
-}
-
-// calcStressScenario は指定の金利・空室率オフセットでシナリオ計算を行い結果を返す
-func calcStressScenario(ctx context.Context, base InvestmentInput, label string, rateDelta, vacDelta float64) StressScenarioResult {
-	_, span := otel.Tracer(domainTracerName).Start(ctx, "domain.calcStressScenario")
-	defer span.End()
-	span.SetAttributes(
-		attribute.String("domain.stress.label", label),
-		attribute.Float64("domain.stress.rate_delta", rateDelta),
-		attribute.Float64("domain.stress.vac_delta", vacDelta),
-	)
-	in := base
-
-	effectiveVacancy := in.VacancyRate + vacDelta
-	if effectiveVacancy > 1 {
-		effectiveVacancy = 1
-	}
-
-	// 初年度賃料（空室率調整済み）— 賃料下落はループ内で年次適用
-	annualRent := in.MonthlyRent * 12 * (1 - effectiveVacancy)
-
-	// 初年度の実効金利（スケジュール+ストレスdelta）
-	initRate := resolveRateForYear(in.AnnualLoanRate, rateDelta, in.RateAdjustmentSchedule, 1)
-	var monthlyPayment float64
-	var monthlyPrincipalStress float64
-	if in.LoanMethod == LoanMethodEqualPrincipal && in.LoanYears > 0 {
-		totalMonths := in.LoanYears * 12
-		monthlyPrincipalStress = in.LoanAmount / float64(totalMonths)
-	} else {
-		monthlyPayment = calcMonthlyPayment(in.LoanAmount, initRate, in.LoanYears)
-	}
-
-	// HoldingYears年間の累積CF（税引後）とブレークイーン年を算出
-	// DSCR は各年返済額から算出した保有期間内最悪値（変動金利上昇ケースで正確なリスク評価を行うため）
-	holdingYears := in.HoldingYears
-	if holdingYears <= 0 {
-		holdingYears = 10
-	}
-	totalCF := 0.0
-	breakEvenYear := -1
-	cumCF := 0.0
-	remainingBalance := in.LoanAmount
-	currentRate := initRate
-	curMonthlyPayment := monthlyPayment
-	minDSCR := math.MaxFloat64
-	hasLoanYear := false
-	for y := 1; y <= holdingYears; y++ {
-		// 変動金利スケジュール適用（元利均等のみ月次返済額を再計算）
-		if y > 1 && len(in.RateAdjustmentSchedule) > 0 {
-			newRate := resolveRateForYear(in.AnnualLoanRate, rateDelta, in.RateAdjustmentSchedule, y)
-			if newRate != currentRate && remainingBalance > 0 && y <= in.LoanYears {
-				if in.LoanMethod != LoanMethodEqualPrincipal {
-					remainingYears := in.LoanYears - (y - 1)
-					curMonthlyPayment = calcMonthlyPayment(remainingBalance, newRate, remainingYears)
-				}
-				currentRate = newRate
-			}
-		}
-
-		yearLoan := 0.0
-		yearInterest := 0.0
-		if remainingBalance > 0 && y <= in.LoanYears {
-			if in.LoanMethod == LoanMethodEqualPrincipal {
-				yi, yp := calcYearlyLoanComponentsEqualPrincipal(remainingBalance, currentRate, monthlyPrincipalStress)
-				yearLoan = yi + yp
-				yearInterest = yi
-				remainingBalance -= yp
-			} else {
-				annInterest, annPrincipal := calcYearlyLoanComponents(remainingBalance, currentRate, curMonthlyPayment)
-				yearLoan = curMonthlyPayment * 12
-				yearInterest = annInterest
-				remainingBalance -= annPrincipal
-			}
-			if remainingBalance < 0 {
-				remainingBalance = 0
-			}
-		}
-
-		// 賃料下落率を年次適用（ストレステストは保守的評価のため RentGrowthRate を意図的に無視）
-		// RentGrowthRate を考慮した楽観シナリオはメインの Analyze() で rentForYear() が担う。
-		declineFactor := math.Pow(1-in.RentDeclineRate, float64(y-1))
-		yearRent := annualRent * declineFactor
-		yearExpenses := yearRent*in.ExpenseRate + in.AnnualPropertyTax
-		yearNOI := yearRent - yearExpenses
-
-		if yearLoan > 0 {
-			hasLoanYear = true
-			if yearDSCR := yearNOI / yearLoan; yearDSCR < minDSCR {
-				minDSCR = yearDSCR
-			}
-		}
-
-		cf := yearNOI - yearLoan
-		// 減価償却は省略した保守的近似（簡略ストレス計算のため過大に税を見積もる）
-		taxableIncome := yearNOI - yearInterest
-		incomeTax := 0.0
-		if taxableIncome > 0 {
-			incomeTax = taxableIncome * in.IncomeTaxRate
-		}
-		afterTaxCF := cf - incomeTax
-		totalCF += afterTaxCF
-		cumCF += afterTaxCF
-		if breakEvenYear == -1 && cumCF > 0 {
-			breakEvenYear = y
-		}
-	}
-	dscr := 0.0
-	if hasLoanYear {
-		dscr = minDSCR
-	}
-
-	// BreakEvenYear は税引後 cumCF が初めて正転した年（#312）
-	// IsSafe の DSCR 閾値は 1.0 のまま維持（UI バッジは別途 1.2 で表示、#313）
-	isSafe := false
-	if !hasLoanYear {
-		// 保有期間内に返済が発生しない場合（無借金物件等）はブレークイーン達成のみで安全と判定
-		isSafe = breakEvenYear != -1 && breakEvenYear <= holdingYears
-	} else {
-		isSafe = dscr >= 1.0 && breakEvenYear != -1 && breakEvenYear <= holdingYears
-	}
-
-	return StressScenarioResult{
-		Label:             label,
-		InterestRateDelta: rateDelta,
-		VacancyRateDelta:  vacDelta,
-		TotalCashFlow:     totalCF,
-		DSCR:              dscr,
-		BreakEvenYear:     breakEvenYear,
-		IsSafe:            isSafe,
-	}
-}
-
 // calcCriticalErrors はPhase 1の重大リスク項目を判定する
 func calcCriticalErrors(input InvestmentInput, deadCrossYear, usefulLife int) []CriticalError {
 	var errs []CriticalError
@@ -643,20 +148,6 @@ func calcCriticalErrors(input InvestmentInput, deadCrossYear, usefulLife int) []
 	return errs
 }
 
-// calcMonthlyPayment は元利均等返済の月次返済額を計算する
-// P: 元金, annualRate: 年利, years: 返済年数
-func calcMonthlyPayment(principal, annualRate float64, years int) float64 {
-	if principal <= 0 || years <= 0 {
-		return 0
-	}
-	if annualRate == 0 {
-		return principal / float64(years*12)
-	}
-	r := annualRate / 12
-	n := float64(years * 12)
-	return principal * r * math.Pow(1+r, n) / (math.Pow(1+r, n) - 1)
-}
-
 // CalcDSCR は借入金償還余裕率を返す（DSCR = NOI / 年間返済額）
 // annualDebtService が 0 以下の場合は 0 を返す。
 func CalcDSCR(noi, annualDebtService float64) float64 {
@@ -664,282 +155,6 @@ func CalcDSCR(noi, annualDebtService float64) float64 {
 		return 0
 	}
 	return noi / annualDebtService
-}
-
-// CalcEqualPrincipalPayment は元金均等返済の month 回目（1始まり）の月次返済額を返す。
-// principal: 借入元本, annualRate: 年利, months: 総返済回数, month: 対象月（1〜months）
-func CalcEqualPrincipalPayment(principal, annualRate float64, months, month int) float64 {
-	if principal <= 0 || months <= 0 || month < 1 || month > months {
-		return 0
-	}
-	monthlyPrincipal := principal / float64(months)
-	remainingBalance := principal - float64(month-1)*monthlyPrincipal
-	monthlyInterest := remainingBalance * (annualRate / 12)
-	return monthlyPrincipal + monthlyInterest
-}
-
-// calcYearlyLoanComponentsEqualPrincipal は元金均等返済における1年分の利息・元金返済額を計算する。
-// monthlyPrincipal: 毎月の固定元金返済額（= 元本 / 総返済回数）
-func calcYearlyLoanComponentsEqualPrincipal(balance, annualRate, monthlyPrincipal float64) (interest, principal float64) {
-	r := annualRate / 12
-	remaining := balance
-	for range 12 {
-		if remaining <= 0 {
-			break
-		}
-		mp := monthlyPrincipal
-		if mp > remaining {
-			mp = remaining
-		}
-		interest += remaining * r
-		principal += mp
-		remaining -= mp
-	}
-	return interest, principal
-}
-
-// CalcLTVSensitivity は指定の LTV 水準ごとに DSCR・CF・CF利回りを算出して返す。
-// ltvRange が nil の場合は [0.5, 0.6, 0.7, 0.8, 0.9] を使用する。
-func CalcLTVSensitivity(input InvestmentInput, ltvRange []float64) []LTVSensitivityRow {
-	input.Defaults()
-	if ltvRange == nil {
-		ltvRange = []float64{0.5, 0.6, 0.7, 0.8, 0.9}
-	}
-
-	miscExpenses := (input.LandPrice + input.BuildingCost) * input.MiscExpenseRate
-	totalInvestment := input.LandPrice + input.BuildingCost + miscExpenses
-	if totalInvestment <= 0 {
-		return nil
-	}
-
-	effectiveVacancy := math.Min(input.VacancyRate, 0.99)
-	annualRent := input.MonthlyRent * 12 * (1 - effectiveVacancy)
-	annualExpenses := annualRent*input.ExpenseRate + input.AnnualPropertyTax
-	noi := annualRent - annualExpenses
-
-	rate1 := resolveRateForYear(input.AnnualLoanRate, input.LoanRateDelta, input.RateAdjustmentSchedule, 1)
-
-	rows := make([]LTVSensitivityRow, 0, len(ltvRange))
-	for _, ltv := range ltvRange {
-		loanAmount := totalInvestment * ltv
-		equity := totalInvestment * (1 - ltv)
-
-		var annualDebtService float64
-		if input.LoanMethod == LoanMethodEqualPrincipal && input.LoanYears > 0 {
-			totalMonths := input.LoanYears * 12
-			monthlyPrincipal := loanAmount / float64(totalMonths)
-			yearInterest, yearPrincipal := calcYearlyLoanComponentsEqualPrincipal(
-				loanAmount, rate1, monthlyPrincipal,
-			)
-			annualDebtService = yearInterest + yearPrincipal
-		} else {
-			mp := calcMonthlyPayment(loanAmount, rate1, input.LoanYears)
-			annualDebtService = mp * 12
-		}
-
-		dscr := CalcDSCR(noi, annualDebtService)
-		annualCF := noi - annualDebtService
-		cfYield := annualCF / totalInvestment
-
-		rows = append(rows, LTVSensitivityRow{
-			LTV:        ltv,
-			Equity:     equity,
-			LoanAmount: loanAmount,
-			DSCR:       dscr,
-			AnnualCF:   annualCF,
-			CFYield:    cfYield,
-		})
-	}
-	return rows
-}
-
-// calcYearlyLoanComponents は1年分の利息・元金返済額を計算する
-// 最終月で月次返済額が残高を超える場合は残高のみを元金返済として扱い、誤差を防ぐ
-func calcYearlyLoanComponents(balance, annualRate, monthlyPayment float64) (interest, principal float64) {
-	if annualRate == 0 {
-		if monthlyPayment*12 > balance {
-			return 0, balance
-		}
-		return 0, monthlyPayment * 12
-	}
-	r := annualRate / 12
-	remaining := balance
-	for range 12 {
-		if remaining <= 0 {
-			break
-		}
-		monthInterest := remaining * r
-		monthPrincipal := monthlyPayment - monthInterest
-		// 最終月: 残高が月次元金返済より少ない場合は残高のみ返済
-		if monthPrincipal > remaining {
-			monthPrincipal = remaining
-		}
-		interest += monthInterest
-		principal += monthPrincipal
-		remaining -= monthPrincipal
-	}
-	return interest, principal
-}
-
-// calcRequiredForTarget は目標利回り達成に必要な値を逆算する
-// costReduction は「土地価格または建築費のいずれか一方」を削減すべき金額を表す
-func calcRequiredForTarget(input InvestmentInput, totalInvestment float64) (requiredRent, costReduction float64) {
-	target := input.YieldTarget
-	requiredAnnualRent := totalInvestment * target
-	requiredRent = requiredAnnualRent / 12
-
-	currentAnnualRent := input.MonthlyRent * 12
-	requiredTotalInvestment := currentAnnualRent / target
-
-	excess := totalInvestment - requiredTotalInvestment
-	if excess > 0 {
-		costReduction = excess
-	}
-	return requiredRent, costReduction
-}
-
-// CalcNPV は将来キャッシュフロー・ターミナルバリュー・割引率・初期投資から NPV を計算する
-func CalcNPV(cfs []float64, terminalValue, discountRate, initialInvestment float64) float64 {
-	pv := 0.0
-	for t, cf := range cfs {
-		pv += cf / math.Pow(1+discountRate, float64(t+1))
-	}
-	n := len(cfs)
-	if n > 0 {
-		pv += terminalValue / math.Pow(1+discountRate, float64(n))
-	}
-	return pv - initialInvestment
-}
-
-// CalcIRR は二分法で IRR を求める。収束しない場合は nil と error を返す。
-func CalcIRR(cfs []float64, terminalValue, initialInvestment float64) (*float64, error) {
-	const (
-		lo      = -0.50
-		hi      = 2.00
-		maxIter = 200
-		tol     = 1.0
-	)
-	npvLo := CalcNPV(cfs, terminalValue, lo, initialInvestment)
-	npvHi := CalcNPV(cfs, terminalValue, hi, initialInvestment)
-	if npvLo*npvHi > 0 {
-		return nil, fmt.Errorf("IRR: no root in [%.0f%%, %.0f%%]", lo*100, hi*100)
-	}
-	low, high := lo, hi
-	for i := 0; i < maxIter; i++ {
-		mid := (low + high) / 2
-		npvMid := CalcNPV(cfs, terminalValue, mid, initialInvestment)
-		if math.Abs(npvMid) < tol {
-			v := mid
-			return &v, nil
-		}
-		if npvMid*npvLo < 0 {
-			high = mid
-		} else {
-			low = mid
-			npvLo = npvMid
-		}
-	}
-	return nil, fmt.Errorf("IRR: did not converge after %d iterations", maxIter)
-}
-
-// calcTerminalValueWithDecline は価格下落率を考慮した売却時ターミナルバリューを計算する
-// 注意: 売却費用・譲渡税の計算ロジックは calcExit と重複している。
-//       税率・仲介手数料率を変更する際は両方を更新すること。
-func calcTerminalValueWithDecline(
-	input InvestmentInput,
-	yearly []YearlyResult,
-	adjustedSalePrice float64,
-	accumulatedDepreciation float64,
-	miscExpenses float64,
-) float64 {
-	holdIdx := input.HoldingYears - 1
-	if holdIdx >= len(yearly) {
-		holdIdx = len(yearly) - 1
-	}
-	exitYear := yearly[holdIdx]
-	sellExpenses := (adjustedSalePrice*0.03+60_000) * 1.10
-	bookValueBuilding := math.Max(input.BuildingCost-accumulatedDepreciation, 0)
-	acquisitionCost := input.LandPrice + bookValueBuilding + miscExpenses
-	capGain := adjustedSalePrice - sellExpenses - acquisitionCost
-	var transferTax float64
-	if capGain > 0 {
-		var taxRate float64
-		if input.HoldingYears > 5 {
-			taxRate = longTermTransferTaxRate
-		} else {
-			taxRate = shortTermTransferTaxRate
-		}
-		transferTax = capGain * taxRate
-	}
-	return adjustedSalePrice - sellExpenses - transferTax - exitYear.RemainingLoanBalance
-}
-
-// 譲渡所得税率（所得税＋復興特別所得税＋住民税）
-// 根拠: 租税特別措置法31条・32条、復興財源確保法33条（2037年まで）
-// 注意: 租税特別措置法31条の3の10年超軽減（14.21%）は「居住用財産」の特例であり
-//       投資用物件には適用されない。投資用は保有年数に関わらず長期=20.315%。
-const (
-	shortTermTransferTaxRate = 0.3963  // 短期（5年以下）: 所得税30%+復興0.63%+住民税9%
-	longTermTransferTaxRate  = 0.20315 // 長期（5年超）: 所得税15%+復興0.315%+住民税5%
-)
-
-// calcExit は出口戦略（売却）の試算を行う
-//
-// 売却価格: NOI（純収益）/ 目標利回り（実質ベース）で収益還元法により算出
-// 取得費: 土地 + 建物簿価 + 取得時諸経費（税法上の取得費）
-// 売却費用: 仲介手数料の上限額を概算控除（消費税込み）
-// 税率: 保有5年超で長期、10年超で軽減税率を適用
-func calcExit(input InvestmentInput, yearly []YearlyResult, accumulatedDepreciation float64, miscExpenses float64) (
-	salePrice, capitalGain, transferTax, netProceeds, totalEquity float64,
-) {
-	if len(yearly) == 0 || input.HoldingYears <= 0 || input.ExitYieldTarget <= 0 {
-		return
-	}
-
-	holdIdx := input.HoldingYears - 1
-	if holdIdx >= len(yearly) {
-		holdIdx = len(yearly) - 1
-	}
-
-	exitYear := yearly[holdIdx]
-
-	// 収益還元法: 売却価格 = NOI / 目標利回り（実質ベース）
-	// NOI = 実効賃料収入 - 運営経費（ローン利息は含まない）
-	noi := exitYear.AnnualRent - exitYear.AnnualExpenses
-	salePrice = noi / input.ExitYieldTarget
-
-	// 売却費用（仲介手数料上限額の概算・消費税込み）
-	// 根拠: 宅建業法46条 上限 = 売却価格×3%+6万円（+消費税10%）
-	sellExpenses := (salePrice*0.03+60_000) * 1.10
-
-	// 建物の税務上の簿価（定額法累計控除後）
-	bookValueBuilding := input.BuildingCost - accumulatedDepreciation
-	if bookValueBuilding < 0 {
-		bookValueBuilding = 0
-	}
-
-	// 取得費 = 土地取得費 + 建物簿価 + 取得時諸経費
-	// 根拠: 所得税法38条（取得費に含まれる付随費用）
-	acquisitionCost := input.LandPrice + bookValueBuilding + miscExpenses
-
-	// 譲渡所得 = 売却価格 - 売却費用 - 取得費
-	capitalGain = salePrice - sellExpenses - acquisitionCost
-
-	if capitalGain > 0 {
-		// 投資用物件の譲渡所得税: 保有5年超=長期(20.315%)、5年以下=短期(39.63%)の2段階
-		// 租税特別措置法31条の3の10年超軽減(14.21%)は居住用財産の特例のため対象外
-		var taxRate float64
-		if input.HoldingYears > 5 {
-			taxRate = longTermTransferTaxRate
-		} else {
-			taxRate = shortTermTransferTaxRate
-		}
-		transferTax = capitalGain * taxRate
-	}
-
-	netProceeds = salePrice - sellExpenses - transferTax - exitYear.RemainingLoanBalance
-	totalEquity = netProceeds + exitYear.CumulativeCashFlow
-	return
 }
 
 // CalcLandPriceStats は取引データから統計を算出する
@@ -997,276 +212,6 @@ func CalcLandPriceStats(ctx context.Context, transactions []LandTransaction) Lan
 		Zoning:         zoning,
 		UrbanRisks:     detectUrbanRisks(transactions, zoning),
 	}
-}
-
-// calcZoningSummary は取引データから最頻の用途地域情報を抽出する
-func calcZoningSummary(transactions []LandTransaction) *ZoningSummary {
-	if len(transactions) == 0 {
-		return nil
-	}
-	cp := modalString(transactions, func(t LandTransaction) string { return t.CityPlanning })
-	bc := modalString(transactions, func(t LandTransaction) string { return t.BuildingCoverage })
-	far := modalString(transactions, func(t LandTransaction) string { return t.FloorAreaRatio })
-	if cp == "" && bc == "" && far == "" {
-		return nil
-	}
-	return &ZoningSummary{
-		CityPlanning:     cp,
-		BuildingCoverage: bc,
-		FloorAreaRatio:   far,
-	}
-}
-
-// detectUrbanRisks は取引データと用途地域サマリーから都市計画リスクを検出する
-func detectUrbanRisks(transactions []LandTransaction, zoning *ZoningSummary) []UrbanRisk {
-	var risks []UrbanRisk
-	if zoning == nil {
-		return risks
-	}
-
-	// 市街化調整区域（最頻値が調整区域）
-	if strings.Contains(zoning.CityPlanning, "市街化調整区域") {
-		risks = append(risks, UrbanRisk{
-			Code:        "URBANIZATION_CONTROL_ZONE",
-			Level:       UrbanRiskLevelError,
-			Title:       "市街化調整区域",
-			Description: "原則として新たな建築・用途変更が制限されます。既存建物の建替えが困難になる可能性があります。",
-		})
-	}
-
-	// 都市計画区域外 / 非線引き区域
-	if strings.Contains(zoning.CityPlanning, "非線引") || zoning.CityPlanning == "都市計画区域外" {
-		risks = append(risks, UrbanRisk{
-			Code:        "UNZONED_AREA",
-			Level:       UrbanRiskLevelWarning,
-			Title:       "非線引き・都市計画区域外",
-			Description: "インフラ整備が遅れやすく、将来の資産価値が不安定になる可能性があります。",
-		})
-	}
-
-	// 調整区域が最頻値でなくとも30%以上混在する場合の注意
-	if !strings.Contains(zoning.CityPlanning, "市街化調整区域") {
-		controlCount, totalWithCP := 0, 0
-		for _, t := range transactions {
-			if t.CityPlanning != "" {
-				totalWithCP++
-				if strings.Contains(t.CityPlanning, "市街化調整区域") {
-					controlCount++
-				}
-			}
-		}
-		if totalWithCP > 0 && float64(controlCount)/float64(totalWithCP) >= 0.3 {
-			ratio := float64(controlCount) / float64(totalWithCP) * 100
-			risks = append(risks, UrbanRisk{
-				Code:        "MIXED_ZONE_CAUTION",
-				Level:       UrbanRiskLevelWarning,
-				Title:       "市街化調整区域が混在",
-				Description: fmt.Sprintf("エリア内取引の約%.0f%%が市街化調整区域です。対象物件の区域区分を必ず確認してください。", ratio),
-			})
-		}
-	}
-
-	return risks
-}
-
-// BuildUrbanRisksFromAPIs は MLIT 専用 API（XKT003/020/030/XST001）の結果からリスクを構築する。
-// detectUrbanRisks（XIT001テキスト検出）とは独立して使用し、結果を呼び出し元でマージする。
-func BuildUrbanRisksFromAPIs(
-	locationItems []LocationOptimizationItem,
-	embankmentItems []EmbankmentItem,
-	roadItems []UrbanRoadItem,
-	disasters []DisasterHistoryItem,
-) []UrbanRisk {
-	var risks []UrbanRisk
-
-	// XKT003: 立地適正化計画フィーチャが存在し、居住誘導区域が含まれない場合
-	if len(locationItems) > 0 {
-		hasResidential := false
-		for _, item := range locationItems {
-			if strings.Contains(item.KubunNameJa, "居住誘導区域") {
-				hasResidential = true
-				break
-			}
-		}
-		if !hasResidential {
-			risks = append(risks, UrbanRisk{
-				Code:        "OUTSIDE_RESIDENTIAL_GUIDANCE",
-				Level:       UrbanRiskLevelWarning,
-				Title:       "居住誘導区域外",
-				Description: "立地適正化計画の居住誘導区域外です。将来的に行政サービスの縮小・インフラ維持コスト増加の可能性があります（コンパクトシティ計画）。",
-			})
-		}
-	}
-
-	// XKT020: 大規模盛土造成地フィーチャが存在する場合
-	if len(embankmentItems) > 0 {
-		embDesc := "大規模盛土造成地に該当します。地震時の沈下・崩壊リスクがあります。"
-		if c := embankmentItems[0].Classification; c != "" {
-			embDesc = fmt.Sprintf("大規模盛土造成地（%s）に該当します。地震時の沈下・崩壊リスクがあります。", c)
-		}
-		risks = append(risks, UrbanRisk{
-			Code:        "LARGE_EMBANKMENT",
-			Level:       UrbanRiskLevelWarning,
-			Title:       "大規模盛土造成地",
-			Description: embDesc,
-		})
-	}
-
-	// XKT030: 都市計画道路（kubun_id=3011）フィーチャが存在する場合
-	for _, item := range roadItems {
-		if item.KubunID == 3011 {
-			risks = append(risks, UrbanRisk{
-				Code:        "URBAN_PLANNING_ROAD",
-				Level:       UrbanRiskLevelWarning,
-				Title:       "都市計画道路の予定地",
-				Description: "都市計画道路の予定地に一部かかっています。将来的に建物の一部または全部が収用対象となる可能性があります。",
-			})
-			break
-		}
-	}
-
-	// XST001: 災害履歴フィーチャが存在する場合
-	if len(disasters) > 0 {
-		names := make([]string, 0, len(disasters))
-		seen := make(map[string]bool)
-		for _, d := range disasters {
-			if d.Name != "" && !seen[d.Name] {
-				names = append(names, d.Name)
-				seen[d.Name] = true
-			}
-		}
-		desc := "このエリアで過去に災害が記録されています。"
-		if len(names) > 0 {
-			desc = fmt.Sprintf("このエリアで過去に災害が記録されています（%s）。", strings.Join(names, "・"))
-		}
-		risks = append(risks, UrbanRisk{
-			Code:        "DISASTER_HISTORY",
-			Level:       UrbanRiskLevelWarning,
-			Title:       "災害履歴あり",
-			Description: desc,
-		})
-	}
-
-	return risks
-}
-
-// BuildHazardRisks は XKT026–029 ハザード API の結果を UrbanRisk スライスに変換する。
-// 各ハザード種別で最も深刻な 1 件のみを返す（重複排除）。
-func BuildHazardRisks(
-	floods []FloodHazardItem,
-	storms []StormHazardItem,
-	tsunamis []TsunamiHazardItem,
-	landslides []LandslideHazardItem,
-) []UrbanRisk {
-	var risks []UrbanRisk
-
-	// XKT026: 洪水浸水想定区域 — DepthRank >= 3 を ERROR、それ以外を WARNING
-	if len(floods) > 0 {
-		worst := floods[0]
-		for _, f := range floods[1:] {
-			if f.DepthRank > worst.DepthRank {
-				worst = f
-			}
-		}
-		level := UrbanRiskLevelWarning
-		if worst.DepthRank >= 3 {
-			level = UrbanRiskLevelError
-		}
-		desc := fmt.Sprintf("洪水浸水想定区域（深さランク %d）に該当します。大雨・河川氾濫時に浸水リスクがあります。", worst.DepthRank)
-		if worst.RiverName != "" {
-			desc = fmt.Sprintf("洪水浸水想定区域（%s、深さランク %d）に該当します。大雨・河川氾濫時に浸水リスクがあります。", worst.RiverName, worst.DepthRank)
-		}
-		risks = append(risks, UrbanRisk{
-			Code:        "FLOOD_HAZARD",
-			Level:       level,
-			Title:       "洪水浸水想定区域",
-			Description: desc,
-		})
-	}
-
-	// XKT027: 高潮浸水想定区域 — 存在すれば WARNING（深さ文字列は最初の非空エントリを採用）
-	if len(storms) > 0 {
-		depthJa := ""
-		for _, s := range storms {
-			if s.DepthJa != "" {
-				depthJa = s.DepthJa
-				break
-			}
-		}
-		desc := "高潮浸水想定区域に該当します。台風・高波時に浸水リスクがあります。"
-		if depthJa != "" {
-			desc = fmt.Sprintf("高潮浸水想定区域（%s）に該当します。台風・高波時に浸水リスクがあります。", depthJa)
-		}
-		risks = append(risks, UrbanRisk{
-			Code:        "STORM_HAZARD",
-			Level:       UrbanRiskLevelWarning,
-			Title:       "高潮浸水想定区域",
-			Description: desc,
-		})
-	}
-
-	// XKT028: 津波浸水想定区域 — 存在すれば ERROR
-	if len(tsunamis) > 0 {
-		desc := "津波浸水想定区域に該当します。沿岸部の津波・高潮による浸水リスクがあります。"
-		if tsunamis[0].DepthJa != "" {
-			desc = fmt.Sprintf("津波浸水想定区域（%s）に該当します。沿岸部の津波・高潮による浸水リスクがあります。", tsunamis[0].DepthJa)
-		}
-		risks = append(risks, UrbanRisk{
-			Code:        "TSUNAMI_HAZARD",
-			Level:       UrbanRiskLevelError,
-			Title:       "津波浸水想定区域",
-			Description: desc,
-		})
-	}
-
-	// XKT029: 土砂災害警戒区域 — ZoneCode=1（特別警戒）を ERROR、ZoneCode=2（警戒）を WARNING
-	if len(landslides) > 0 {
-		// 最も深刻な ZoneCode（1 < 2、つまり 1 が最重要）を選択
-		worst := landslides[0]
-		for _, l := range landslides[1:] {
-			if l.ZoneCode < worst.ZoneCode {
-				worst = l
-			}
-		}
-		level := UrbanRiskLevelWarning
-		zoneName := "警戒区域"
-		if worst.ZoneCode == 1 {
-			level = UrbanRiskLevelError
-			zoneName = "特別警戒区域"
-		}
-		phenomenonNames := map[int]string{1: "急傾斜地崩壊", 2: "土石流", 3: "地すべり"}
-		pName, ok := phenomenonNames[worst.PhenomenonType]
-		if !ok {
-			pName = "土砂災害"
-		}
-		desc := fmt.Sprintf("土砂災害%s（%s）に該当します。大雨時の土砂崩れ・土石流リスクがあります。", zoneName, pName)
-		risks = append(risks, UrbanRisk{
-			Code:        "LANDSLIDE_HAZARD",
-			Level:       level,
-			Title:       "土砂災害警戒区域",
-			Description: desc,
-		})
-	}
-
-	return risks
-}
-
-// modalString は transactions から getter で取得した文字列の最頻値を返す（空文字は除外）
-func modalString(transactions []LandTransaction, getter func(LandTransaction) string) string {
-	counts := make(map[string]int, len(transactions))
-	for _, t := range transactions {
-		v := getter(t)
-		if v != "" {
-			counts[v]++
-		}
-	}
-	best, bestCount := "", 0
-	for v, c := range counts {
-		if c > bestCount {
-			best, bestCount = v, c
-		}
-	}
-	return best
 }
 
 // CompareLandPrice は検討中の土地価格と相場を比較する
